commit provisorio

+++ src/components/generic/ImageUploadPreview.jsx
@@ -1,43 +1,84 @@
-import { useEffect, useState } from "react";
-import { Button, Form, Image } from "react-bootstrap";
-import { AiFillDelete } from "react-icons/ai";
+import React, { useEffect, useState } from "react";
+import { Button, Form } from "react-bootstrap";
 
+import { t } from "@lingui/macro";
+import { useDispatch, useSelector } from "react-redux";
+import {
+  openMediaLibrary,
+  openUploadWidget,
+} from "../../services/CloudinaryService";
+import { selectLoading, setLoading } from "../../slices/genericSlice";
+import CustomSpinner from "./CustomSpinner";
 import "./ImageUploadPreview.scss";
 
-function ImageUploadPreview({ item }) {
-  const { images } = item;
+export const ImageUploadPreview = React.forwardRef((props, ref) => {
+  const dispatch = useDispatch();
+  const { images, isAddMode, imgId, setValue, errors, idItem } = props;
+  const loading = useSelector(selectLoading);
+
   const [files, setFiles] = useState([]);
+  const [filesObj, setFilesObj] = useState([]);
 
   useEffect(() => {
-
     if (images) {
       setFiles(images);
     }
   }, [images]);
 
-  const uploadFiles = (e) => {
-    e.preventDefault();
+  const uploadImageWithCloudinary = () => {
+    dispatch(setLoading({ loading: true }));
+
+    openUploadWidget("Posts", idItem, () => {
+      dispatch(setLoading({ loading: false }));
+    });
   };
+
+  const mediaLibraryCloudinary = () => {
+    const mediaLibraryOptions = {
+      search: { expression: `tags=Posts/201` },
+    };
+    openMediaLibrary(mediaLibraryOptions);
+  };
+
   const uploadMultipleFiles = (e) => {
     const _files = Object.assign([], files);
+    const _filesObj = Object.assign([], filesObj);
 
     for (let i = 0; i < e.target.files.length; i++) {
       _files.push(URL.createObjectURL(e.target.files[i]));
+      _filesObj.push(e.target.files[i]);
     }
 
     setFiles(_files);
+    setFilesObj(_filesObj);
+    setValue("images", _filesObj);
   };
 
   function deleteFile(e) {
     const s = files.filter((item, index) => index !== e);
+    const o = filesObj.filter((item, index) => index !== e);
     setFiles(s);
+    setFilesObj(o);
+    setValue("images", o);
   }
 
+  const open = () => {
+    openUploadWidget();
+  };
+
+  if (loading) {
+    return <CustomSpinner />;
+  }
+
+  const imagesFunction = isAddMode
+    ? uploadImageWithCloudinary
+    : mediaLibraryCloudinary;
+
   return (
-    <Form>
-      <Form.Group style={{ display: "flex" , flexWrap:"wrap"}}>
+    <div>
+      {/* <Form.Group style={{ display: "flex", flexWrap: "wrap" }}>
         {(files || []).map((url, idx) => (
-          <div  key={idx} className="container-upload-preview">
+          <div key={idx} className="container-upload-preview">
             <AiFillDelete onClick={() => deleteFile(idx)} />
             <Image
               className="upload-preview"
@@ -48,26 +89,20 @@ function ImageUploadPreview({ item }) {
             />
           </div>
         ))}
-      </Form.Group>
+      </Form.Group> */}
 
-      <Form.Group style={{ display: "flex" }}>
-        <input
+      <Form.Group>
+        <Button onClick={imagesFunction}> {t`new.item.images`} </Button>
+
+        {/* <input
           type="file"
-          className="form-control"
+          className={`form-control ${errors?.images ? "is-invalid" : ""}`}
           accept="image/*"
           onChange={uploadMultipleFiles}
           multiple
-        />
-        <Button
-          type="button"
-          className="upload-button btn btn-danger btn-block"
-          onClick={uploadFiles}
-        >
-          Upload
-        </Button>
+        /> */}
+        <div className="invalid-feedback">{errors?.images?.message}</div>
       </Form.Group>
-    </Form>
+    </div>
   );
-}
-
-export default ImageUploadPreview;
+});
